add code

+++ wm/src/zidl/window_proxy.cpp
@@ -462,7 +462,7 @@ WMError WindowProxy::NotifyScreenshotAppEvent(ScreenshotEventType type)
         TLOGE(WmsLogTag::WMS_ATTRIBUTE, "WriteInterfaceToken failed");
         return WMError::WM_ERROR_IPC_FAILED;
     }
-    if (!data.WriteInt32(type)) {
+    if (!data.WriteInt32(static_cast<int32_t>(type))) {
         TLOGE(WmsLogTag::WMS_ATTRIBUTE, "Write screenshot event type failed");
         return WMError::WM_ERROR_IPC_FAILED;
     }
